requirements/base.txt
# Core dependencies
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.2
Flask-Mail==0.10.0
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.36
alembic==1.13.1

# Security
werkzeug==3.1.3
bcrypt==4.2.1
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2
cryptography==42.0.5

# Forms and validation
WTForms==3.2.1
email-validator==2.2.0

# Data processing
pandas==2.3.0
numpy==2.3.1
openpyxl==3.1.5
holidays==0.63

# PDF processing
PyPDF2==3.0.1
pdf2image==1.17.0
camelot-py[cv]==0.11.0

# Image processing
pillow==11.3.0
pytesseract==0.3.13
opencv-python==4.10.0.84

# Optimization
PuLP==2.7.0

# Utilities
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
click==8.2.1

# Redis for caching/sessions
redis==5.0.1